refactor: 유저에게 입력 받는 부분은 User 클래스로 이동
